# Swagger-docs.yaml
---
swagger: "2.0"
info:
  description: |
    This is the Swagger 2.0 API for Web Architectures course group project work.
    Group "1"
  version: 1.1.1
  title: Make me a sandwich
  contact:
    email: duy.a.vu@tuni.fi
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost
basePath: /v1
tags:
- name: order
  description: Sandwich orders
- name: sandwich
  description: |
    NOT MANDATORY TO IMPLEMENT!
    Types of sandwich that users can order- CRUD
- name: user
  description: |
    NOT MANDATORY TO IMPLEMENT!
    User management
schemes:
- http
paths:
  /order:
    get:
      tags:
      - order
      summary: Get a list of all orders. Empty array if no orders are found.
      operationId: getOrders
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/ArrayOfOrders'
    post:
      tags:
      - order
      summary: Add an order for an sandwich
      operationId: addOrder
      produces:
      - application/json
      parameters:
      - in: body
        name: order
        description: place an order for a sandwich
        required: true
        schema:
          $ref: '#/definitions/Order'
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        "400":
          description: Order not created
  /order/{orderId}:
    get:
      tags:
      - order
      summary: Find an order by its ID
      description: IDs must be positive integers
      operationId: getOrderById
      produces:
      - application/json
      parameters:
      - name: orderId
        in: path
        description: ID of the order that needs to be fetched
        required: true
        type: integer
        minimum: 1
        format: int64
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
  /sandwich:
    get:
      tags:
      - sandwich
      summary: Get a list of all sandwiches. Empty array if no sandwiches are found.
      operationId: getSandwiches
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/ArrayOfSandwiches'
    post:
      tags:
      - sandwich
      summary: Add a new sandwich to the store. Needs an API key.
      operationId: addSandwich
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Sandwich object that needs to be added to the store
        required: true
        schema:
          $ref: '#/definitions/Sandwich'
      responses:
        "200":
          description: Sandwich created
          schema:
            $ref: '#/definitions/Sandwich'
        "401":
          description: Invalid API key
        "405":
          description: Invalid input
      security:
      - api_key: []
  /sandwich/{sandwichId}:
    get:
      tags:
      - sandwich
      summary: Find sandwich by ID
      description: Returns a single sandwich
      operationId: getSandwichById
      produces:
      - application/json
      parameters:
      - name: sandwichId
        in: path
        description: ID of sandwich to return
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Sandwich'
        "400":
          description: Invalid ID supplied
        "404":
          description: Sandwich not found
    put:
      tags:
      - sandwich
      summary: Updates a sandwich in the store with JSON in body. Needs an API key
      operationId: updateSandwich
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: sandwichId
        in: path
        description: ID of sandwich to return
        required: true
        type: integer
        format: int64
      - in: body
        name: body
        description: Sandwich object that needs to be added to the store
        required: true
        schema:
          $ref: '#/definitions/Sandwich'
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Sandwich'
        "400":
          description: Invalid ID supplied
        "401":
          description: Invalid API key
        "405":
          description: Invalid input
      security:
      - api_key: []
    delete:
      tags:
      - sandwich
      summary: Deletes a sandwich. Needs an API key
      operationId: deleteSandwich
      produces:
      - application/json
      parameters:
      - name: sandwichId
        in: path
        description: Sandwich id to delete
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid ID supplied
        "401":
          description: Invalid API key
        "404":
          description: Sandwich not found
      security:
      - api_key: []
  /user:
    post:
      tags:
      - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Created user object
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        "201":
          description: successful operation
          schema:
            $ref: '#/definitions/UserWithoutPassword'
        "400":
          description: invalid input
      security:
      - api_key: []
  /user/login:
    post:
      tags:
      - user
      summary: Logs user into the system
      operationId: loginUser
      produces:
      - application/json
      parameters:
      - in: body
        name: user
        description: The user for login
        required: true
        schema:
          $ref: '#/definitions/UsernameWithPassword'
      responses:
        "200":
          description: successful operation. API key is returned
          schema:
            type: string
        "400":
          description: Invalid username/password supplied
  /user/logout:
    post:
      tags:
      - user
      summary: Logs out current logged in user session
      operationId: logoutUser
      produces:
      - application/json
      parameters:
      - in: body
        name: username
        description: Object containing the username of user to be logged out
        required: true
        schema:
          type: object
          properties:
            username:
              type: string
      responses:
        "200":
          description: successful operation. Logged out user's API key is deleted
  /user/{username}:
    get:
      tags:
      - user
      summary: Get user by user name
      operationId: getUserByName
      produces:
      - application/json
      parameters:
      - name: username
        in: path
        description: The name that needs to be fetched. Use user1 for testing.
        required: true
        type: string
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/UserWithoutPassword'
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
    put:
      tags:
      - user
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      produces:
      - application/json
      parameters:
      - name: username
        in: path
        description: name that need to be updated
        required: true
        type: string
      - in: body
        name: body
        description: Updated user object
        required: true
        schema:
          $ref: '#/definitions/User'
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/UserWithoutPassword'
        "400":
          description: Invalid user supplied
        "401":
          description: Invalid API key
        "404":
          description: User not found
      security:
      - api_key: []
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      produces:
      - application/json
      parameters:
      - name: username
        in: path
        description: The name that needs to be deleted
        required: true
        type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid username supplied
        "401":
          description: Invalid API key
        "404":
          description: User not found
      security:
      - api_key: []
  /user/apikey/{username}:
    get:
      tags:
      - user
      summary: Get latest apikey. This is used when one user is logged in from multiple clients
      operationId: getLatestAPIkey
      produces:
      - application/text
      parameters:
      - name: username
        in: path
        description: username whose API key needs to be updated
        required: true
        type: string
      responses:
        "200":
          description: successful operation. New API key is returned
          schema:
            type: string
        "401":
          description: User not logged in
securityDefinitions:
  sandwichestore_auth:
    type: oauth2
    authorizationUrl: http://sandwichestore.swagger.io/oauth/dialog
    flow: implicit
    scopes:
      write:sandwiches: modify sandwiches in your account
      read:sandwiches: read your sandwiches
  api_key:
    type: apiKey
    name: X-API-KEY
    in: header
definitions:
  Order:
    type: object
    properties:
      id:
        type: integer
        format: int64
      sandwichId:
        type: integer
        format: int64
      status:
        type: string
        description: Order Status
        enum:
        - ordered
        - received
        - inQueue
        - ready
        - failed
    xml:
      name: Order
  ArrayOfOrders:
    type: array
    items:
      $ref: '#/definitions/Order'
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      email:
        type: string
      password:
        type: string
    xml:
      name: User
  UserWithoutPassword:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      email:
        type: string
    xml:
      name: UserWithoutPassword
  UsernameWithPassword:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    xml:
      name: UsernameWithPassword
  Topping:
    type: object
    required:
    - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Tag
  Sandwich:
    type: object
    required:
    - breadType
    - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        example: my-super-sandwich
      toppings:
        type: array
        xml:
          name: toppings
          wrapped: true
        items:
          $ref: '#/definitions/Topping'
      breadType:
        type: string
        description: bread used in the sandwich
        enum:
        - oat
        - rye
        - wheat
    xml:
      name: Sandwich
  ArrayOfSandwiches:
    type: array
    items:
      $ref: '#/definitions/Sandwich'